composer/2.1: Add check for display existence in ComposerClient
Add check for display existence in ComposerClient. It is possible
for ComposerClient to receive display disconnect from HWC, when it
is in the middle of creating or destroying layers based on calls
from SF. ComposerClient immediately destroys its copy of display
but doesn't check if display exists in subsequent creation,
destruction of layers.

Change-Id: I50ed8a4394320d35cc0ac152bec98b9cc3f8423b
CRs-fixed: 2149612